// todo-webapp/src/generated/todo-api.ts
export interface components {
    schemas: {
        TodoItem: {
            id: string;
            title: string;
            completed: boolean;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
        CreateTodoItem: {
            title: string;
        };
        UpdateTodoItem: {
            title?: string;
            completed?: boolean;
        };
        Error: {
            /** @description HTTP or application error code */
            code: number;
            /** @description short human-readable label */
            message: string;
            /** @description detailed explanation */
            description?: string;
            /** @description URI to documentation */
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}